fix: move common types
