+++ apps/webapp/types/routing.type.ts
@@ -9,3 +9,10 @@ export interface CommonPageParams extends LangProp {}
 export interface CommonPageProps {
   params: CommonPageParams
 }
+
+export type ProjectPageParams = { project: string; lang: Lang }
+
+export type ProjectPageProps = {
+  children: React.ReactNode
+  params: ProjectPageParams
+}
